+++ cmd/esbuild/main.go
@@ -77,7 +77,9 @@ var helpText = func(colors logger.Colors) string {
   --sourcemap=inline        Emit the source map with an inline data URL
   --tree-shaking=...        Set to "ignore-annotations" to work with packages
                             that have incorrect tree-shaking annotations
-  --tsconfig=...            Use this tsconfig.json file instead of other ones
+	--tsconfig=...            Use this tsconfig.json file instead of other ones
+	--remove--console         Removes console calls
+	--remove--debugger        Removes debugger statements
   --version                 Print the current version and exit (` + esbuildVersion + `)
 
 ` + colors.Bold + `Examples:` + colors.Default + `