+++ lib/common.ts
@@ -106,6 +106,8 @@ function pushCommonFlags(flags: string[], options: CommonOptions, keys: OptionKe
   let keepNames = getFlag(options, keys, 'keepNames', mustBeBoolean);
   let banner = getFlag(options, keys, 'banner', mustBeString);
   let footer = getFlag(options, keys, 'footer', mustBeString);
+  let removeConsole = getFlag(options, keys, 'removeConsole', mustBeBoolean);
+  let removeDebugger = getFlag(options, keys, 'removeDebugger', mustBeBoolean);
 
   if (target) {
     if (Array.isArray(target)) flags.push(`--target=${Array.from(target).map(validateTarget).join(',')}`)
@@ -135,6 +137,9 @@ function pushCommonFlags(flags: string[], options: CommonOptions, keys: OptionKe
 
   if (banner) flags.push(`--banner=${banner}`);
   if (footer) flags.push(`--footer=${footer}`);
+
+  if (removeConsole) flags.push('--remove-console');
+  if (removeDebugger) flags.push('--remove-debugger');
 }
 
 function flagsForBuildOptions(callName: string, options: types.BuildOptions, isTTY: boolean, logLevelDefault: types.LogLevel, writeDefault: boolean):