+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Proposal
+
+* Add `--remove-console` option that avoids print console calls to target
+* Add `--remove-debugger` option that avoids print debugger statement to target
+
 ## 0.8.24
 
 * Share reference-counted service instances internally ([#600](https://github.com/evanw/esbuild/issues/600))